Batch of 64-bit fixes.  Removes _M_IX86 from being hardcoded in cmake as well.

+++ codemp/game/g_main.c
@@ -59,7 +59,7 @@ This must be the very first function compiled into the .q3vm file
 		G_ShutdownGame( arg0 );
 		return 0;
 	case GAME_CLIENT_CONNECT:
-		return (int)ClientConnect( arg0, arg1, arg2 );
+		return (intptr_t)ClientConnect( arg0, arg1, arg2 );
 	case GAME_CLIENT_THINK:
 		ClientThink( arg0, NULL );
 		return 0;
